logic to handle when owner should leave if it is not first in queue
Subscribes to all other vehicle statuses, and takes the first vehicle in queue's status to see if it will move or will be on standby for longer. This is so the current vehicle can decide whether to let owner leave or wait to move up to the front before the owner leaves. This is all in order to avoid congestion at the drop off zone.